pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["Mike Ebinum <mebinum@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
run = "scraper.main:main"
flush = "scraper.database:flush_db"

[tool.poetry.dependencies]
python = "3.11.5"
psycopg2 = "^2.9.9"
splinter = "^0.21.0"
sqlalchemy = { version = "2.0.28", extras = ["postgresql"] }
python-dotenv = "^1.0.1"
requests = "^2.31.0"
selenium = "^4.18.1"
pandas = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
mypy = "^1.4.1"
ruff = "^0.2.1"
pytest-unordered = "^0.5.2"